Handling older worker version for environment reload request.
