--- host/src/FunctionsNetHost/Grpc/UnsupportedVersionException.cs
@@ -1,12 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the MIT License. See License.txt in the project root for license information.
-
-namespace Microsoft.Azure.Functions.Worker
-{
-    public class UnsupportedVersionException : Exception
-    {
-        public UnsupportedVersionException() { }
-
-        public UnsupportedVersionException(string message) : base(message) { }
-    }
-}
